Integrate development changes. (#4)
* Testing gitlab push / CI.

* Testing gitlab push / CI.

* Implemented player stats (data-layer & log).

* Implemented event (player) models.

* Implemented separate player component for mobile stub.

* Implemented (player) events queries & admin.

* Added vscode config.

* Added vscode config.

* Implemented mobile player stub.

* Refactored queue & player states & controls to be more re-usable.

* Cleaned up button components.

* Updated GCP library versions.

* Fixed airplay sync time handling.

* Refactored stats & history handling.

* Cleaned up code to pass linting / style.

* Implemented handling of (in-)active editors.

* Implemented handling of (in-)active editors.

* Switched cloud-build to cloud-builders/gcloud.

* Implemented filter- & tag style changes.

* Implementing context-menu (tracks / media).

* Implemented extended queue handling (insert next / end).

* Implemented static / "cms" content.

* Refactored "Topbar" to "Navigation"

* Implemented context-menu blueprints.

* Changed series number to INT.

* Refactored "load more" handling.

* Implemented dynamic context-menu positioning.

* Implemented ISRC as (media) identifier.

* Implemented ISRC as (media) identifier.

* Implemented consistent "order by" (latest airplay / latest emission) in discover.

* Changed playlist table to show duration instead of num. emissions.

* Implemented account migration blueprint.

* Implemented account migration flow.

* Implemented platform account (one-way) sync.

* Refactored scheduling logic to not use 'node-scheduler'.

* Fixed play (on-air) component.

* Integrating layout & styling changes.

* Refactored scss to use native variables.

* Implemented dynamic color theme for main navigation.

* Configured gitlab repository mirror. (#3)

+++ core/catalog/models/playlist.py
@@ -35,7 +35,7 @@ class Playlist(
         on_delete=models.SET_NULL,
     )
 
-    series_episode = models.PositiveSmallIntegerField(
+    series_episode = models.PositiveIntegerField(
         verbose_name="Series #",
         null=True,
         blank=True,
